Make use of 'launcher::test_server' and 'launcher::test_database' for integration tests
