+++ eth-bytecode-db/eth-bytecode-db-server/Cargo.toml
@@ -32,6 +32,7 @@ smart-contract-verifier-proto = { git = "https://github.com/blockscout/blockscou
 verifier-alliance-entity = { path = "../eth-bytecode-db/verifier-alliance-entity" }
 verifier-alliance-migration = { path = "../eth-bytecode-db/verifier-alliance-migration" }
 
+blockscout-service-launcher = { version = "*", features = [ "test-server", "test-database" ] }
 bytes = "1.5.0"
 hex = "0.4.3"
 mockall = "0.11"
